Finalize first pass of registerPlugin API
